refactor: Move func in notifications/utils in helpers file
nothing else has changed
