+++ src/ducks/notifications/CategoryBudget/index.js
@@ -7,7 +7,7 @@ import { Q } from 'cozy-client'
 
 import { ACCOUNT_DOCTYPE, GROUP_DOCTYPE } from 'doctypes'
 import NotificationView from 'ducks/notifications/BaseNotificationView'
-import { formatAmount, getCurrentDate } from 'ducks/notifications/utils'
+import { formatAmount, getCurrentDate } from 'ducks/notifications/helpers'
 import { getCategoryName } from 'ducks/categories/categoriesMap'
 import { getGroupLabel } from 'ducks/groups/helpers'
 import { getAccountLabel } from 'ducks/account/helpers'
