+++ src/ducks/notifications/HealthBillLinked/index.js
@@ -1,18 +1,19 @@
-import NotificationView from 'ducks/notifications/BaseNotificationView'
-import { isHealthExpense } from 'ducks/categories/helpers'
-import { Bill } from 'models'
 import keyBy from 'lodash/keyBy'
 import merge from 'lodash/merge'
 
 import log from 'cozy-logger'
+import { toText } from 'cozy-notifications'
+
+import { Bill } from 'models'
 import {
   treatedByFormat,
   prepareTransactions,
-  getCurrentDate
-} from 'ducks/notifications/utils'
-import { getReimbursementBillIds } from 'ducks/notifications/helpers'
+  getCurrentDate,
+  getReimbursementBillIds
+} from 'ducks/notifications/helpers'
 import template from './template.hbs'
-import { toText } from 'cozy-notifications'
+import NotificationView from 'ducks/notifications/BaseNotificationView'
+import { isHealthExpense } from 'ducks/categories/helpers'
 
 const ACCOUNT_SEL = '.js-account'
 const DATE_SEL = '.js-date'
